requirements.txt
openai>=1.0
pydantic
typer
rich
python-dotenv
